requirements.txt
streamlit>=1.30.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
lxml>=4.9.0
apscheduler>=3.10.0
reportlab>=4.0.0
plotly>=5.18.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
